requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2022.12.7 # via requests
chardet==3.0.4  # via requests
idna==2.7   # via requests
requests==2.20.0    # via -r requirements.in
six==1.10.0 # via websocket-client
urllib3==1.24.3 # via requests
websocket-client==0.32.0
python-dotenv==1.0.0
influxdb3-python==0.8.0
